% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_GeomPedigreePoint.R
\name{Scales}
\alias{Scales}
\alias{scale_feature.name_discrete}
\alias{scale_feature.name}
\alias{scale_sex_discrete}
\alias{scale_sex}
\alias{scale_isdead_discrete}
\alias{scale_isdead}
\title{Scale constructors}
\usage{
scale_feature.name_discrete(
  ...,
  set = "Set2",
  na.value = "grey50",
  main.feature.black = TRUE,
  name = "Features"
)

scale_feature.name(
  ...,
  set = "Set2",
  na.value = "grey50",
  main.feature.black = TRUE,
  name = "Features"
)

scale_sex_discrete(
  ...,
  values = c(22, 22, 22, 22, 21, 21, 21, 21, 0),
  breaks = c("1", "M", "Male", "male", "2", "F", "Female", "female", "3"),
  name = "Sex"
)

scale_sex(
  ...,
  values = c(22, 22, 22, 22, 21, 21, 21, 21, 0),
  breaks = c("1", "M", "Male", "male", "2", "F", "Female", "female", "3"),
  name = "Sex"
)

scale_isdead_discrete(
  ...,
  values = c(0, 0, 47, 47, 63, 63, 63),
  breaks = c("0", FALSE, "1", TRUE, "2", "Unknown", "NA"),
  name = "Status"
)

scale_isdead(
  ...,
  values = c(0, 0, 47, 47, 63, 63, 63),
  breaks = c("0", FALSE, "1", TRUE, "2", "Unknown", "NA"),
  name = "Status"
)
}
\arguments{
\item{...}{Additional parameters passed on to \link{ggplot2:discrete_scale}.}

\item{set}{A palette name from the lists in \link{RColorBrewer:brewer.pal}.}

\item{na.value}{What aesthetic value should the missing values be displayed as?}

\item{main.feature.black}{Should the main (first) feature always be displayed as black?}

\item{name}{Legend title.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
}
\description{
Discrete scale constructors for \link{geom_pedigreepoint}
}
\section{Functions}{
\itemize{
\item \code{scale_feature.name_discrete}: Scale constructor for the \var{feature.name} argument.

\item \code{scale_feature.name}: Alias for \code{scale_feature.name_discrete}.

\item \code{scale_sex_discrete}: Scale constructor for the \var{sex} argument.

\item \code{scale_sex}: Alias for \code{scale_sex_discrete}.

\item \code{scale_isdead_discrete}: Scale constructor for the \var{isdead} argument.

\item \code{scale_isdead}: Alias for \code{scale_isdead_discrete}.
}}

