% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_feature.R
\name{split_by_feature}
\alias{split_by_feature}
\title{split_by_feature}
\usage{
split_by_feature(vector, features, factor)
}
\arguments{
\item{vector}{A vector containing data for drawing GeomPedigreePoint}

\item{features}{Features to split along}

\item{factor}{A scale factor (i.e. desired point size)}
}
\description{
Helper function to split coordinate vector for Pedigree Point generation.
}
